+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.28.0] - 2025-03-06
+
+* runner: Add `Partitioner` and `Runner::with_partitioner` to enable partitioning of test cases, allowing only a subset of the glob result to be executed. This can be helpful for running tests in parallel in CI.
+* bin: Add `--partition-id` and `--partition-count` to set a hash partitioning for the test cases. If users are running in Buildkite CI with `parallelism: ..` specified in the workflow file, this will be automatically configured.
+
 ## [0.27.2] - 2025-02-18
 
 * engines/bin: fix stdin to be closed properly to avoid hangs in the `external` engine.